refactor: distribute api test fixtures via pytest entry-point plugin
- Ship api/tests/plugin.py as a pytest11 entry point so the api's
  --ci option and fixture set are auto-discovered by any venv with
  flagsmith-api[dev] installed.
- Drop py-modules = ["conftest", "manage"]; api/conftest.py is gone
  (its hooks moved into the plugin).
- Fixtures stay in api/tests/fixtures.py and get lazily registered
  in pytest_configure (after pytest-django has set up Django).
